move wr_buf_ functions to a separate file
