+++ src/cases/var_wr_I_case.cpp
@@ -119,114 +119,13 @@
     if (rec_no == 0) nvars_D[dp]++;                                       \
 }
 
-/*----< wr_buf_init() >------------------------------------------------------*/
-void wr_buf_init(io_buffers &buf,
-                 int         gap)
-{
-    buf.gap             = gap;
-
-    buf.fix_txt_buflen  = 0;
-    buf.fix_int_buflen  = 0;
-    buf.fix_dbl_buflen  = 0;
-    buf.fix_buflen      = 0;
-    buf.rec_txt_buflen  = 0;
-    buf.rec_int_buflen  = 0;
-    buf.rec_dbl_buflen  = 0;
-    buf.rec_buflen      = 0;
-
-    buf.fix_txt_buf  = NULL;
-    buf.fix_int_buf  = NULL;
-    buf.fix_dbl_buf  = NULL;
-    buf.fix_buf      = NULL;
-    buf.rec_txt_buf  = NULL;
-    buf.rec_int_buf  = NULL;
-    buf.rec_dbl_buf  = NULL;
-    buf.rec_buf      = NULL;
-}
-
-/*----< wr_buf_malloc() >----------------------------------------------------*/
-int wr_buf_malloc(e3sm_io_config &cfg,
-                  int             one_flush,
-                  io_buffers     &buf)
-{
-    int rank;
-    size_t j;
-
-    MPI_Comm_rank(cfg.io_comm, &rank);
-
-    if (one_flush && cfg.api == pnetcdf) {
-        /* write buffers should not be touched when using PnetCDF iput before
-         * ncmpi_wait_all is called. For HDF5 and ADIOS blob I/O, write data
-         * will be copied and cached into internally allocated buffers and user
-         * buffers can be reused after put call returned.
-         */
-        buf.rec_dbl_buflen *= cfg.nrecs;
-        buf.rec_int_buflen *= cfg.nrecs;
-        buf.rec_txt_buflen *= cfg.nrecs;
-        buf.rec_buflen     *= cfg.nrecs;
-    }
-
-    /* allocate and initialize write buffers */
-    buf.fix_buf     = (vtype*)  malloc(buf.fix_buflen     * sizeof(vtype));
-    buf.fix_txt_buf = (char*)   malloc(buf.fix_txt_buflen * sizeof(char));
-    buf.fix_int_buf = (int*)    malloc(buf.fix_int_buflen * sizeof(int));
-    buf.fix_dbl_buf = (double*) malloc(buf.fix_dbl_buflen * sizeof(double));
-    buf.rec_txt_buf = (char*)   malloc(buf.rec_txt_buflen * sizeof(char));
-    buf.rec_int_buf = (int*)    malloc(buf.rec_int_buflen * sizeof(int));
-    buf.rec_dbl_buf = (double*) malloc(buf.rec_dbl_buflen * sizeof(double));
-    buf.rec_buf     = (vtype*)  malloc(buf.rec_buflen     * sizeof(vtype));
-
-    for (j=0; j<buf.fix_txt_buflen; j++) buf.fix_txt_buf[j] = 'a' + rank;
-    for (j=0; j<buf.fix_int_buflen; j++) buf.fix_int_buf[j] = rank;
-    for (j=0; j<buf.fix_dbl_buflen; j++) buf.fix_dbl_buf[j] = rank;
-    for (j=0; j<buf.fix_buflen;     j++) buf.fix_buf[j]     = rank;
-    for (j=0; j<buf.rec_txt_buflen; j++) buf.rec_txt_buf[j] = 'a' + rank;
-    for (j=0; j<buf.rec_int_buflen; j++) buf.rec_int_buf[j] = rank;
-    for (j=0; j<buf.rec_dbl_buflen; j++) buf.rec_dbl_buf[j] = rank;
-    for (j=0; j<buf.rec_buflen;     j++) buf.rec_buf[j]     = rank;
-
-    return 0;
-}
-
-/*----< wr_buf_free() >------------------------------------------------------*/
-void wr_buf_free(io_buffers &buf)
-{
-    if (buf.fix_txt_buf  != NULL) free(buf.fix_txt_buf);
-    if (buf.fix_int_buf  != NULL) free(buf.fix_int_buf);
-    if (buf.fix_dbl_buf  != NULL) free(buf.fix_dbl_buf);
-    if (buf.fix_buf      != NULL) free(buf.fix_buf);
-    if (buf.rec_txt_buf  != NULL) free(buf.rec_txt_buf);
-    if (buf.rec_int_buf  != NULL) free(buf.rec_int_buf);
-    if (buf.rec_dbl_buf  != NULL) free(buf.rec_dbl_buf);
-    if (buf.rec_buf      != NULL) free(buf.rec_buf);
-
-    buf.fix_txt_buf  = NULL;
-    buf.fix_int_buf  = NULL;
-    buf.fix_dbl_buf  = NULL;
-    buf.fix_buf      = NULL;
-    buf.rec_txt_buf  = NULL;
-    buf.rec_int_buf  = NULL;
-    buf.rec_dbl_buf  = NULL;
-    buf.rec_buf      = NULL;
-
-    buf.fix_txt_buflen  = 0;
-    buf.fix_int_buflen  = 0;
-    buf.fix_dbl_buflen  = 0;
-    buf.fix_buflen      = 0;
-    buf.rec_txt_buflen  = 0;
-    buf.rec_int_buflen  = 0;
-    buf.rec_dbl_buflen  = 0;
-    buf.rec_buflen      = 0;
-}
-
 /*----< var_wr_I_case() >----------------------------------------------------*/
 int var_wr_I_case(e3sm_io_config &cfg,
                   e3sm_io_decom  &decom,
-                  e3sm_io_driver &driver)
+                  e3sm_io_driver &driver,
+                  char           *outfile)
 {
-    char outfile[1040], base_name[1024], *ext;
     char *fix_txt_buf_ptr, *rec_txt_buf_ptr;
-    const char *hist;
     int i, j, err, sub_rank, global_rank, ncid=-1, nflushes=0, one_flush;
     int rec_no, gap=0, my_nreqs, num_decomp_vars;
     int contig_nreqs[MAX_NUM_DECOMP], *nvars_D=cfg.nvars_D;
@@ -267,27 +166,6 @@ int var_wr_I_case(e3sm_io_config &cfg,
         return -1;
     }
 #endif
-
-    /* construct h0/h1 file name */
-    hist = (cfg.nvars == 560) ? "_h0" :  "_h1";
-    ext = strrchr(cfg.out_path, '.');
-    if (ext == NULL || (strcmp(ext, ".nc") && strcmp(ext, ".h5"))) {
-        sprintf(base_name, "%s%s", cfg.out_path, hist);
-        if (cfg.strategy == blob) /* set output subfile name */
-            sprintf(outfile, "%s%s.%04d", cfg.out_path, hist, cfg.subfile_ID);
-        else
-            sprintf(outfile, "%s%s", cfg.out_path, hist);
-    }
-    else {
-        sprintf(outfile, "%s", cfg.out_path);
-        if (cfg.strategy == blob) /* set output subfile name */
-            sprintf(outfile + (ext - cfg.out_path), "%s%s.%04d", hist, ext,
-                    cfg.subfile_ID);
-        else
-            sprintf(outfile + (ext - cfg.out_path), "%s%s", hist, ext);
-        sprintf(base_name, "%s", cfg.out_path);
-        sprintf(base_name + (ext - cfg.out_path), "%s%s", hist, ext);
-    }
 
     if (cfg.verbose) {
         if (cfg.strategy == blob && sub_rank == 0)
@@ -521,9 +399,6 @@ int var_wr_I_case(e3sm_io_config &cfg,
 
     /* check if there is any PnetCDF internal malloc residue */
     check_malloc(&cfg, &driver);
-
-    /* report timing breakdowns */
-    report_timing_WR(&cfg, &driver, &decom, base_name);
 
     /* print MPI-IO hints actually used */
     if (cfg.verbose && global_rank == 0) print_info(&info_used);
